+++ app/controllers/helper/faq_page_controller.rb
@@ -1,29 +1,15 @@
 module Helper
     class FaqPageController < ApplicationController
-        before_action :set_faq_page
-
-        def edit
-            authorize :helper, :manage_faq?
-        end
-
-        def update
-            authorize :helper, :manage_faq?
-
-            if @faq_page.update(faq_page_params)
-                redirect_to edit_helper_faq_page_path, notice: "Successfully updated."
-            else
-                render :edit, status: :unprocessable_entity
-            end
-        end
+        include FaqPageManagement
 
         private
 
-        def set_faq_page
-            @faq_page = FaqPage.instance
+        def authorize_faq!
+            authorize :helper, :manage_faq?
         end
 
-        def faq_page_params
-            params.require(:faq_page).permit(:body)
+        def faq_edit_path
+            edit_helper_faq_page_path
         end
     end
 end
